add build commands and organize react package
